複数プレイを想定してフォームで情報を保持するようにしました。

+++ src/main/java/com/example/quiz/service/QuizService.java
@@ -25,6 +25,6 @@ public interface QuizService {
 	 */
 	
 	/** 登録されているクイズをカウントする */
-	Long getCountQuestion();
+	int getCountQuestion();
 	
 }
